@@ -2,35 +2,26 @@ var fs = require('fs');
 
 // abstract writing screen shot to a file
 function writeScreenShot(data, filename) {
-    var stream = fs.createWriteStream(filename);
-    stream.write(new Buffer(data, 'base64'));
+    var stream = fs.createWriteStream(filename+'.png');
+    stream.write(new Buffer.from(data, 'base64'));
     stream.end();
+    // var stream = fs.createWriteStream(filename+'.txt');
+    // stream.write(data);
+    // stream.end();
+    // console.log(data);
 }
 
 describe('index resource page', function () {
-    it('login to app', function () {
-        browser.waitForAngularEnabled(true);
-        browser.get(browser.baseUrl + '/dist/index.html#/login');
-        browser.sleep(500);
-        browser.takeScreenshot().then(function (png) {
-            writeScreenShot(png, 'test/e2e/screenshots/login_to_app_' + browser.browserName + '.png');
-        });
-        element(by.id('login')).sendKeys('bblier');
-        browser.sleep(500);
-        element(by.id('password')).sendKeys('maarch');
-        browser.sleep(500);
-        element(by.id('submit')).click();
-    });
-
     it('index a document whitout file', function () {
-        browser.sleep(1000);
+        browser.sleep(2000);
+        expect(browser.getCurrentUrl()).toEqual(browser.baseUrl + "/dist/index.html#/home");
         browser.takeScreenshot().then(function (png) {
-            writeScreenShot(png, 'test/e2e/screenshots/home_' + browser.browserName + '.png');
+            writeScreenShot(png, 'test/e2e/screenshots/home_' + browser.browserName);
         });
         element(by.id('indexing')).click();
         browser.sleep(500);
         browser.takeScreenshot().then(function (png) {
-            writeScreenShot(png, 'test/e2e/screenshots/index_a_document_' + browser.browserName + '.png');
+            writeScreenShot(png, 'test/e2e/screenshots/index_a_document_' + browser.browserName);
         });
         element(by.id('doctype')).click();
         browser.sleep(500);
