@@ -79,7 +79,7 @@ if ($xmlconfig == false) {
 
 // Load config
 $config          = $xmlconfig->CONFIG;
-$maarchDirectory = (string)$config->MaarchDirectory;
+$maarchDirectory = (string)$config->maarchDirectory;
 $customID        = (string)$config->customID;
 $customIDPath    = '';
 
@@ -89,7 +89,7 @@ if ($customID <> '') {
 
 chdir($maarchDirectory);
 
-$maarchUrl = (string)$config->MaarchUrl;
+$maarchUrl = (string)$config->maarchUrl;
 
 $GLOBALS['customId']  = $customID;
 $GLOBALS['batchDirectory'] = $maarchDirectory . 'bin'
