@@ -74,7 +74,7 @@ class ContactGroupController
             return $response->withStatus(400)->withJson(['errors' => _CONTACTS_GROUP_LABEL_ALREADY_EXISTS]);
         }
 
-        if ($GLOBALS['login'] = 'superadmin') {
+        if ($GLOBALS['login'] == 'superadmin') {
             $entityOwner = 'superadmin';
         } else {
             $primaryEntity = UserModel::getPrimaryEntityById(['id' => $GLOBALS['id'], 'select' => ['entities.entity_id']]);
