@@ -15,6 +15,7 @@
 namespace Alfresco\controllers;
 
 use Attachment\models\AttachmentModel;
+use Configuration\models\ConfigurationModel;
 use Convert\controllers\ConvertPdfController;
 use Docserver\models\DocserverModel;
 use Entity\models\EntityModel;
@@ -31,54 +32,276 @@ use User\models\UserModel;
 
 class AlfrescoController
 {
+    public function getConfiguration(Request $request, Response $response)
+    {
+        if (!PrivilegeController::hasPrivilege(['privilegeId' => 'admin_alfresco', 'userId' => $GLOBALS['id']])) {
+            return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
+        }
+
+        $configuration = ConfigurationModel::getByService(['service' => 'admin_alfresco']);
+        if (empty($configuration)) {
+            return $response->withJson(['configuration' => null]);
+        }
+
+        $configuration['value'] = json_decode($configuration['value'], true);
+
+        return $response->withJson(['configuration' => ['uri' => $configuration['value']['uri']]]);
+    }
+
+    public function updateConfiguration(Request $request, Response $response)
+    {
+        if (!PrivilegeController::hasPrivilege(['privilegeId' => 'admin_alfresco', 'userId' => $GLOBALS['id']])) {
+            return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
+        }
+
+        $body = $request->getParsedBody();
+
+        if (!Validator::stringType()->notEmpty()->validate($body['uri'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Body uri is empty or not a string']);
+        }
+
+        $value = json_encode(['uri' => $body['uri']]);
+
+        $configuration = ConfigurationModel::getByService(['service' => 'admin_alfresco']);
+        if (empty($configuration)) {
+            ConfigurationModel::create(['service' => 'admin_alfresco', 'value' => $value]);
+        } else {
+            ConfigurationModel::update(['set' => ['value' => $value], 'where' => ['service = ?'], 'data' => ['admin_alfresco']]);
+        }
+
+        return $response->withStatus(204);
+    }
+
+    public function getAvailableEntities(Request $request, Response $response)
+    {
+        if (!PrivilegeController::hasPrivilege(['privilegeId' => 'admin_alfresco', 'userId' => $GLOBALS['id']])) {
+            return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
+        }
+
+        $entities = EntityModel::get(['select' => ['id'], 'where' => ["external_id->>'alfresco' is null"]]);
+
+        $availableEntities = array_column($entities, 'id');
+
+        return $response->withJson(['availableEntities' => $availableEntities]);
+    }
+
     public function getAccounts(Request $request, Response $response)
     {
-//        if (!PrivilegeController::hasPrivilege(['privilegeId' => 'admin_alfresco', 'userId' => $GLOBALS['id']])) {
-//            return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
-//        }
+        if (!PrivilegeController::hasPrivilege(['privilegeId' => 'admin_alfresco', 'userId' => $GLOBALS['id']])) {
+            return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
+        }
 
-        $entities = EntityModel::get(['select' => ['external_id'], 'where' => ["external_id->>'alfresco' is not null"]]);
+        $entities = EntityModel::get(['select' => ['external_id', 'short_label'], 'where' => ["external_id->>'alfresco' is not null"]]);
 
         $accounts = [];
         $alreadyAdded = [];
         foreach ($entities as $entity) {
             $alfresco = json_decode($entity['external_id'], true);
-            if (!in_array($alfresco['alfresco']['login'], $alreadyAdded)) {
+            if (!in_array($alfresco['alfresco']['id'], $alreadyAdded)) {
                 $accounts[] = [
-                    'label' => $alfresco['alfresco']['label'],
-                    'login' => $alfresco['alfresco']['login']
+                    'id'            => $alfresco['alfresco']['id'],
+                    'label'         => $alfresco['alfresco']['label'],
+                    'login'         => $alfresco['alfresco']['login'],
+                    'entitiesLabel' => [$entity['short_label']]
                 ];
-                $alreadyAdded[] = $alfresco['alfresco']['login'];
+                $alreadyAdded[] = $alfresco['alfresco']['id'];
+            } else {
+                foreach ($accounts as $key => $value) {
+                    if ($value['id'] == $alfresco['alfresco']['id']) {
+                        $accounts[$key]['entitiesLabel'][] = $entity['short_label'];
+                    }
+                }
             }
         }
 
         return $response->withJson(['accounts' => $accounts]);
     }
 
+    public function getAccountById(Request $request, Response $response, array $args)
+    {
+        if (!PrivilegeController::hasPrivilege(['privilegeId' => 'admin_alfresco', 'userId' => $GLOBALS['id']])) {
+            return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
+        }
+
+        $entities = EntityModel::get(['select' => ['external_id', 'id'], 'where' => ["external_id->'alfresco'->>'id' = ?"], 'data' => [$args['id']]]);
+        if (empty($entities[0])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Account not found']);
+        }
+
+        $alfresco = json_decode($entities[0]['external_id'], true);
+        $account = [
+            'id'        => $alfresco['alfresco']['id'],
+            'label'     => $alfresco['alfresco']['label'],
+            'login'     => $alfresco['alfresco']['login'],
+            'nodeId'    => $alfresco['alfresco']['nodeId'],
+            'entities'  => []
+        ];
+
+        foreach ($entities as $entity) {
+            $account['entities'][] = $entity['id'];
+        }
+
+        return $response->withJson($account);
+    }
+
+    public function createAccount(Request $request, Response $response)
+    {
+        if (!PrivilegeController::hasPrivilege(['privilegeId' => 'admin_alfresco', 'userId' => $GLOBALS['id']])) {
+            return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
+        }
+
+        $body = $request->getParsedBody();
+
+        if (!Validator::stringType()->notEmpty()->validate($body['label'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Body label is empty or not a string']);
+        } elseif (!Validator::stringType()->notEmpty()->validate($body['login'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Body login is empty or not a string']);
+        } elseif (!Validator::stringType()->notEmpty()->validate($body['password'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Body password is empty or not a string']);
+        } elseif (!Validator::stringType()->notEmpty()->validate($body['nodeId'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Body nodeId is empty or not a string']);
+        } elseif (!Validator::arrayType()->notEmpty()->validate($body['entities'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Body entities is empty or not an array']);
+        }
+
+        foreach ($body['entities'] as $entity) {
+            if (!Validator::intVal()->notEmpty()->validate($entity)) {
+                return $response->withStatus(400)->withJson(['errors' => 'Body entities contains no integer values']);
+            }
+        }
+        $entities = EntityModel::get(['select' => ['id'], 'where' => ['id in (?)'], 'data' => [$body['entities']]]);
+        if (count($entities) != count($body['entities'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Some entities do not exist']);
+        }
+
+        $id = CoreConfigModel::uniqueId();
+        $account = [
+            'id'        => $id,
+            'label'     => $body['label'],
+            'login'     => $body['login'],
+            'password'  => PasswordModel::encrypt(['password' => $body['password']]),
+            'nodeId'    => $body['nodeId']
+        ];
+        $account = json_encode($account);
+
+        EntityModel::update([
+            'postSet'   => ['external_id' => "jsonb_set(external_id, '{alfresco}', '{$account}')"],
+            'where'     => ['id in (?)'],
+            'data'      => [$body['entities']]
+        ]);
+
+        return $response->withStatus(204);
+    }
+
+    public function updateAccount(Request $request, Response $response, array $args)
+    {
+        if (!PrivilegeController::hasPrivilege(['privilegeId' => 'admin_alfresco', 'userId' => $GLOBALS['id']])) {
+            return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
+        }
+
+        $body = $request->getParsedBody();
+
+        if (!Validator::stringType()->notEmpty()->validate($body['label'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Body label is empty or not a string']);
+        } elseif (!Validator::stringType()->notEmpty()->validate($body['login'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Body login is empty or not a string']);
+        } elseif (!Validator::stringType()->notEmpty()->validate($body['nodeId'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Body nodeId is empty or not a string']);
+        } elseif (!Validator::arrayType()->notEmpty()->validate($body['entities'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Body entities is empty or not an array']);
+        }
+
+        $accounts = EntityModel::get(['select' => ['external_id', 'id'], 'where' => ["external_id->'alfresco'->>'id' = ?"], 'data' => [$args['id']]]);
+        if (empty($accounts[0])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Account not found']);
+        }
+
+        foreach ($body['entities'] as $entity) {
+            if (!Validator::intVal()->notEmpty()->validate($entity)) {
+                return $response->withStatus(400)->withJson(['errors' => 'Body entities contains no integer values']);
+            }
+        }
+        $entities = EntityModel::get(['select' => ['id'], 'where' => ['id in (?)'], 'data' => [$body['entities']]]);
+        if (count($entities) != count($body['entities'])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Some entities do not exist']);
+        }
+
+        $alfresco = json_decode($accounts[0]['external_id'], true);
+        $account = [
+            'id'        => $args['id'],
+            'label'     => $body['label'],
+            'login'     => $body['login'],
+            'password'  => empty($body['password']) ? $alfresco['alfresco']['password'] : PasswordModel::encrypt(['password' => $body['password']]),
+            'nodeId'    => $body['nodeId']
+        ];
+        $account = json_encode($account);
+
+        EntityModel::update([
+            'postSet'   => ['external_id' => "jsonb_set(external_id, '{alfresco}', '{$account}')"],
+            'where'     => ['id in (?)'],
+            'data'      => [$body['entities']]
+        ]);
+
+        $previousEntities = array_column($accounts, 'id');
+        $entitiesToRemove = array_diff($previousEntities, $body['entities']);
+        if (!empty($entitiesToRemove)) {
+            EntityModel::update([
+                'postSet'   => ['external_id' => "external_id - 'alfresco'"],
+                'where'     => ['id in (?)'],
+                'data'      => [$entitiesToRemove]
+            ]);
+        }
+
+        return $response->withStatus(204);
+    }
+
+    public function deleteAccount(Request $request, Response $response, array $args)
+    {
+        if (!PrivilegeController::hasPrivilege(['privilegeId' => 'admin_alfresco', 'userId' => $GLOBALS['id']])) {
+            return $response->withStatus(403)->withJson(['errors' => 'Service forbidden']);
+        }
+
+        $accounts = EntityModel::get(['select' => ['external_id', 'id'], 'where' => ["external_id->'alfresco'->>'id' = ?"], 'data' => [$args['id']]]);
+        if (empty($accounts[0])) {
+            return $response->withStatus(400)->withJson(['errors' => 'Account not found']);
+        }
+
+        $entitiesToRemove = array_column($accounts, 'id');
+        EntityModel::update([
+            'postSet'   => ['external_id' => "external_id - 'alfresco'"],
+            'where'     => ['id in (?)'],
+            'data'      => [$entitiesToRemove]
+        ]);
+
+        return $response->withStatus(204);
+    }
+
     public function getRootFolders(Request $request, Response $response)
     {
-        $loadedXml = CoreConfigModel::getXmlLoaded(['path' => 'apps/maarch_entreprise/xml/alfrescoConfig.xml']);
-
-        if (empty($loadedXml) || (string)$loadedXml->ENABLED != 'true') {
+        $configuration = ConfigurationModel::getByService(['service' => 'admin_alfresco']);
+        if (empty($configuration)) {
             return $response->withStatus(400)->withJson(['errors' => 'Alfresco configuration is not enabled']);
-        } elseif (empty((string)$loadedXml->URI)) {
+        }
+
+        $configuration = json_decode($configuration['value'], true);
+        if (empty($configuration['uri'])) {
             return $response->withStatus(400)->withJson(['errors' => 'Alfresco configuration URI is empty']);
         }
-        $alfrescoUri = rtrim((string)$loadedXml->URI, '/');
+        $alfrescoUri = rtrim($configuration['uri'], '/');
 
         $entity = UserModel::getPrimaryEntityById(['id' => $GLOBALS['id'], 'select' => ['entities.external_id']]);
         if (empty($entity)) {
             return $response->withStatus(400)->withJson(['errors' => 'User has no primary entity']);
         }
         $entityInformations = json_decode($entity['external_id'], true);
-        if (empty($entityInformations['alfrescoNodeId']) || empty($entityInformations['alfrescoLogin']) || empty($entityInformations['alfrescoPassword'])) {
+        if (empty($entityInformations['alfresco'])) {
             return $response->withStatus(400)->withJson(['errors' => 'User primary entity has not enough alfresco informations']);
         }
-        $entityInformations['alfrescoPassword'] = PasswordModel::decrypt(['cryptedPassword' => $entityInformations['alfrescoPassword']]);
+        $entityInformations['alfresco']['password'] = PasswordModel::decrypt(['cryptedPassword' => $entityInformations['alfresco']['password']]);
 
         $curlResponse = CurlModel::execSimple([
-            'url'           => "{$alfrescoUri}/alfresco/versions/1/nodes/{$entityInformations['alfrescoNodeId']}/children",
-            'basicAuth'     => ['user' => $entityInformations['alfrescoLogin'], 'password' => $entityInformations['alfrescoPassword']],
+            'url'           => "{$alfrescoUri}/alfresco/versions/1/nodes/{$entityInformations['alfresco']['nodeId']}/children",
+            'basicAuth'     => ['user' => $entityInformations['alfresco']['login'], 'password' => $entityInformations['alfresco']['password']],
             'headers'       => ['content-type:application/json'],
             'method'        => 'GET',
             'queryParams'   => ['where' => '(isFolder=true)']
@@ -105,28 +328,30 @@ class AlfrescoController
 
     public function getChildrenFoldersById(Request $request, Response $response, array $args)
     {
-        $loadedXml = CoreConfigModel::getXmlLoaded(['path' => 'apps/maarch_entreprise/xml/alfrescoConfig.xml']);
-
-        if (empty($loadedXml) || (string)$loadedXml->ENABLED != 'true') {
+        $configuration = ConfigurationModel::getByService(['service' => 'admin_alfresco']);
+        if (empty($configuration)) {
             return $response->withStatus(400)->withJson(['errors' => 'Alfresco configuration is not enabled']);
-        } elseif (empty((string)$loadedXml->URI)) {
+        }
+
+        $configuration = json_decode($configuration['value'], true);
+        if (empty($configuration['uri'])) {
             return $response->withStatus(400)->withJson(['errors' => 'Alfresco configuration URI is empty']);
         }
-        $alfrescoUri = rtrim((string)$loadedXml->URI, '/');
+        $alfrescoUri = rtrim($configuration['uri'], '/');
 
         $entity = UserModel::getPrimaryEntityById(['id' => $GLOBALS['id'], 'select' => ['entities.external_id']]);
         if (empty($entity)) {
             return $response->withStatus(400)->withJson(['errors' => 'User has no primary entity']);
         }
         $entityInformations = json_decode($entity['external_id'], true);
-        if (empty($entityInformations['alfrescoNodeId']) || empty($entityInformations['alfrescoLogin']) || empty($entityInformations['alfrescoPassword'])) {
+        if (empty($entityInformations['alfresco'])) {
             return $response->withStatus(400)->withJson(['errors' => 'User primary entity has not enough alfresco informations']);
         }
-        $entityInformations['alfrescoPassword'] = PasswordModel::decrypt(['cryptedPassword' => $entityInformations['alfrescoPassword']]);
+        $entityInformations['alfresco']['password'] = PasswordModel::decrypt(['cryptedPassword' => $entityInformations['alfresco']['password']]);
 
         $curlResponse = CurlModel::execSimple([
             'url'           => "{$alfrescoUri}/alfresco/versions/1/nodes/{$args['id']}/children",
-            'basicAuth'     => ['user' => $entityInformations['alfrescoLogin'], 'password' => $entityInformations['alfrescoPassword']],
+            'basicAuth'     => ['user' => $entityInformations['alfresco']['login'], 'password' => $entityInformations['alfresco']['password']],
             'headers'       => ['content-type:application/json'],
             'method'        => 'GET',
             'queryParams'   => ['where' => '(isFolder=true)']
@@ -160,36 +385,38 @@ class AlfrescoController
             return $response->withStatus(400)->withJson(['errors' => 'Query params search is too short']);
         }
 
-        $loadedXml = CoreConfigModel::getXmlLoaded(['path' => 'apps/maarch_entreprise/xml/alfrescoConfig.xml']);
-
-        if (empty($loadedXml) || (string)$loadedXml->ENABLED != 'true') {
+        $configuration = ConfigurationModel::getByService(['service' => 'admin_alfresco']);
+        if (empty($configuration)) {
             return $response->withStatus(400)->withJson(['errors' => 'Alfresco configuration is not enabled']);
-        } elseif (empty((string)$loadedXml->URI)) {
+        }
+
+        $configuration = json_decode($configuration['value'], true);
+        if (empty($configuration['uri'])) {
             return $response->withStatus(400)->withJson(['errors' => 'Alfresco configuration URI is empty']);
         }
-        $alfrescoUri = rtrim((string)$loadedXml->URI, '/');
+        $alfrescoUri = rtrim($configuration['uri'], '/');
 
         $entity = UserModel::getPrimaryEntityById(['id' => $GLOBALS['id'], 'select' => ['entities.external_id']]);
         if (empty($entity)) {
             return $response->withStatus(400)->withJson(['errors' => 'User has no primary entity']);
         }
         $entityInformations = json_decode($entity['external_id'], true);
-        if (empty($entityInformations['alfrescoNodeId']) || empty($entityInformations['alfrescoLogin']) || empty($entityInformations['alfrescoPassword'])) {
+        if (empty($entityInformations['alfresco'])) {
             return $response->withStatus(400)->withJson(['errors' => 'User primary entity has not enough alfresco informations']);
         }
-        $entityInformations['alfrescoPassword'] = PasswordModel::decrypt(['cryptedPassword' => $entityInformations['alfrescoPassword']]);
+        $entityInformations['alfresco']['password'] = PasswordModel::decrypt(['cryptedPassword' => $entityInformations['alfresco']['password']]);
 
         $search = addslashes($queryParams['search']);
         $body = [
             'query' => [
-                'query'     => "select * from cmis:folder where CONTAINS ('cmis:name:*{$search}*') and IN_TREE('{$entityInformations['alfrescoNodeId']}')",
+                'query'     => "select * from cmis:folder where CONTAINS ('cmis:name:*{$search}*') and IN_TREE('{$entityInformations['alfresco']['nodeId']}')",
                 'language'  => 'cmis',
             ],
             'fields' => ['id', 'name']
         ];
         $curlResponse = CurlModel::execSimple([
             'url'           => "{$alfrescoUri}/search/versions/1/search",
-            'basicAuth'     => ['user' => $entityInformations['alfrescoLogin'], 'password' => $entityInformations['alfrescoPassword']],
+            'basicAuth'     => ['user' => $entityInformations['alfresco']['login'], 'password' => $entityInformations['alfresco']['password']],
             'headers'       => ['content-type:application/json', 'Accept: application/json'],
             'method'        => 'POST',
             'body'          => json_encode($body)
@@ -220,24 +447,26 @@ class AlfrescoController
         ValidatorModel::intVal($args, ['resId', 'userId']);
         ValidatorModel::stringType($args, ['folderId', 'folderName']);
 
-        $loadedXml = CoreConfigModel::getXmlLoaded(['path' => 'apps/maarch_entreprise/xml/alfrescoConfig.xml']);
-
-        if (empty($loadedXml) || (string)$loadedXml->ENABLED != 'true') {
+        $configuration = ConfigurationModel::getByService(['service' => 'admin_alfresco']);
+        if (empty($configuration)) {
             return ['errors' => 'Alfresco configuration is not enabled'];
-        } elseif (empty((string)$loadedXml->URI)) {
+        }
+
+        $configuration = json_decode($configuration['value'], true);
+        if (empty($configuration['uri'])) {
             return ['errors' => 'Alfresco configuration URI is empty'];
         }
-        $alfrescoUri = rtrim((string)$loadedXml->URI, '/');
+        $alfrescoUri = rtrim($configuration['uri'], '/');
 
         $entity = UserModel::getPrimaryEntityById(['id' => $args['userId'], 'select' => ['entities.external_id']]);
         if (empty($entity)) {
             return ['errors' => 'User has no primary entity'];
         }
         $entityInformations = json_decode($entity['external_id'], true);
-        if (empty($entityInformations['alfrescoNodeId']) || empty($entityInformations['alfrescoLogin']) || empty($entityInformations['alfrescoPassword'])) {
+        if (empty($entityInformations['alfresco'])) {
             return ['errors' => 'User primary entity has not enough alfresco informations'];
         }
-        $entityInformations['alfrescoPassword'] = PasswordModel::decrypt(['cryptedPassword' => $entityInformations['alfrescoPassword']]);
+        $entityInformations['alfresco']['password'] = PasswordModel::decrypt(['cryptedPassword' => $entityInformations['alfresco']['password']]);
 
         $document = ResModel::getById(['select' => ['filename', 'subject', 'alt_identifier', 'external_id'], 'resId' => $args['resId']]);
         if (empty($document)) {
@@ -268,7 +497,7 @@ class AlfrescoController
 
         $curlResponse = CurlModel::execSimple([
             'url'           => "{$alfrescoUri}/alfresco/versions/1/nodes/{$args['folderId']}/children",
-            'basicAuth'     => ['user' => $entityInformations['alfrescoLogin'], 'password' => $entityInformations['alfrescoPassword']],
+            'basicAuth'     => ['user' => $entityInformations['alfresco']['login'], 'password' => $entityInformations['alfresco']['password']],
             'headers'       => ['content-type:application/json', 'Accept: application/json'],
             'method'        => 'POST',
             'body'          => json_encode(['name' => str_replace('/', '_', $document['alt_identifier']), 'nodeType' => 'cm:folder'])
@@ -278,12 +507,13 @@ class AlfrescoController
         }
         $resourceFolderId = $curlResponse['response']['entry']['id'];
 
+        $document['subject'] = str_replace([':', '*', '\'', '"', '>', '<'], ' ', $document['subject']);
         $multipartBody = [
             'filedata' => ['isFile' => true, 'filename' => $document['subject'], 'content' => $fileContent],
         ];
         $curlResponse = CurlModel::execSimple([
             'url'           => "{$alfrescoUri}/alfresco/versions/1/nodes/{$resourceFolderId}/children",
-            'basicAuth'     => ['user' => $entityInformations['alfrescoLogin'], 'password' => $entityInformations['alfrescoPassword']],
+            'basicAuth'     => ['user' => $entityInformations['alfresco']['login'], 'password' => $entityInformations['alfresco']['password']],
             'method'        => 'POST',
             'multipartBody' => $multipartBody
         ]);
@@ -299,7 +529,7 @@ class AlfrescoController
         ];
         $curlResponse = CurlModel::execSimple([
             'url'           => "{$alfrescoUri}/alfresco/versions/1/nodes/{$documentId}",
-            'basicAuth'     => ['user' => $entityInformations['alfrescoLogin'], 'password' => $entityInformations['alfrescoPassword']],
+            'basicAuth'     => ['user' => $entityInformations['alfresco']['login'], 'password' => $entityInformations['alfresco']['password']],
             'headers'       => ['content-type:application/json', 'Accept: application/json'],
             'method'        => 'PUT',
             'body'          => json_encode($body)
@@ -348,7 +578,7 @@ class AlfrescoController
             if ($firstAttachment) {
                 $curlResponse = CurlModel::execSimple([
                     'url'           => "{$alfrescoUri}/alfresco/versions/1/nodes/{$resourceFolderId}/children",
-                    'basicAuth'     => ['user' => $entityInformations['alfrescoLogin'], 'password' => $entityInformations['alfrescoPassword']],
+                    'basicAuth'     => ['user' => $entityInformations['alfresco']['login'], 'password' => $entityInformations['alfresco']['password']],
                     'headers'       => ['content-type:application/json', 'Accept: application/json'],
                     'method'        => 'POST',
                     'body'          => json_encode(['name' => 'Pièces jointes', 'nodeType' => 'cm:folder'])
@@ -377,7 +607,7 @@ class AlfrescoController
             ];
             $curlResponse = CurlModel::execSimple([
                 'url'           => "{$alfrescoUri}/alfresco/versions/1/nodes/{$attachmentsFolderId}/children",
-                'basicAuth'     => ['user' => $entityInformations['alfrescoLogin'], 'password' => $entityInformations['alfrescoPassword']],
+                'basicAuth'     => ['user' => $entityInformations['alfresco']['login'], 'password' => $entityInformations['alfresco']['password']],
                 'method'        => 'POST',
                 'multipartBody' => $multipartBody
             ]);
@@ -394,7 +624,7 @@ class AlfrescoController
             ];
             $curlResponse = CurlModel::execSimple([
                 'url'           => "{$alfrescoUri}/alfresco/versions/1/nodes/{$attachmentId}",
-                'basicAuth'     => ['user' => $entityInformations['alfrescoLogin'], 'password' => $entityInformations['alfrescoPassword']],
+                'basicAuth'     => ['user' => $entityInformations['alfresco']['login'], 'password' => $entityInformations['alfresco']['password']],
                 'headers'       => ['content-type:application/json', 'Accept: application/json'],
                 'method'        => 'PUT',
                 'body'          => json_encode($body)
