@@ -1,51 +1,53 @@
-import { NgModule }     from '@angular/core';
+import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-import { AdministrationComponent }                      from './home/administration.component';
-import { UsersAdministrationComponent }                 from './user/users-administration.component';
-import { UserAdministrationComponent }                  from './user/user-administration.component';
-import { GroupsAdministrationComponent }                from './group/groups-administration.component';
-import { GroupAdministrationComponent }                 from './group/group-administration.component';
-import { BasketsAdministrationComponent }               from './basket/baskets-administration.component';
-import { BasketAdministrationComponent }                from './basket/basket-administration.component';
-import { DoctypesAdministrationComponent }              from './doctype/doctypes-administration.component';
-import { DiffusionModelsAdministrationComponent }       from './diffusionModel/diffusionModels-administration.component';
-import { DiffusionModelAdministrationComponent }        from './diffusionModel/diffusionModel-administration.component';
-import { EntitiesAdministrationComponent }              from './entity/entities-administration.component';
-import { StatusesAdministrationComponent }              from './status/statuses-administration.component';
-import { StatusAdministrationComponent }                from './status/status-administration.component';
-import { ActionsAdministrationComponent }               from './action/actions-administration.component';
-import { ActionAdministrationComponent }                from './action/action-administration.component';
-import { ParameterAdministrationComponent }             from './parameter/parameter-administration.component';
-import { ParametersAdministrationComponent }            from './parameter/parameters-administration.component';
-import { PrioritiesAdministrationComponent }            from './priority/priorities-administration.component';
-import { PriorityAdministrationComponent }              from './priority/priority-administration.component';
-import { NotificationsAdministrationComponent }         from './notification/notifications-administration.component';
-import { NotificationAdministrationComponent }          from './notification/notification-administration.component';
-import { HistoryAdministrationComponent }               from './history/history-administration.component';
-import { HistoryBatchAdministrationComponent }          from './history/batch/history-batch-administration.component';
-import { UpdateStatusAdministrationComponent }          from './updateStatus/update-status-administration.component';
-import { ContactsGroupsAdministrationComponent }        from './contact/group/contacts-groups-administration.component';
-import { ContactsGroupAdministrationComponent }         from './contact/group/contacts-group-administration.component';
-import { ContactsParametersAdministrationComponent }       from './contact/parameter/contacts-parameters-administration.component';
-import { VersionsUpdateAdministrationComponent }        from './versionUpdate/versions-update-administration.component';
-import { DocserversAdministrationComponent }            from './docserver/docservers-administration.component';
-import { DocserverAdministrationComponent }             from './docserver/docserver-administration.component';
-import { TemplatesAdministrationComponent }             from './template/templates-administration.component';
-import { TemplateAdministrationComponent }              from './template/template-administration.component';
-import { SecuritiesAdministrationComponent }            from './security/securities-administration.component';
-import { SendmailAdministrationComponent }              from './sendmail/sendmail-administration.component';
-import { ShippingsAdministrationComponent }             from './shipping/shippings-administration.component';
-import { ShippingAdministrationComponent }              from './shipping/shipping-administration.component';
-import { CustomFieldsAdministrationComponent }          from './customField/custom-fields-administration.component';
-import { AppGuard }                                     from '../../service/app.guard';
-import { IndexingModelAdministrationComponent }        from './indexingModel/indexing-model-administration.component';
-import { IndexingModelsAdministrationComponent }        from './indexingModel/indexing-models-administration.component';
-import { ContactsListAdministrationComponent }        from './contact/list/contacts-list-administration.component';
+import { AdministrationComponent } from './home/administration.component';
+import { UsersAdministrationComponent } from './user/users-administration.component';
+import { UserAdministrationComponent } from './user/user-administration.component';
+import { GroupsAdministrationComponent } from './group/groups-administration.component';
+import { GroupAdministrationComponent } from './group/group-administration.component';
+import { BasketsAdministrationComponent } from './basket/baskets-administration.component';
+import { BasketAdministrationComponent } from './basket/basket-administration.component';
+import { DoctypesAdministrationComponent } from './doctype/doctypes-administration.component';
+import { DiffusionModelsAdministrationComponent } from './diffusionModel/diffusionModels-administration.component';
+import { DiffusionModelAdministrationComponent } from './diffusionModel/diffusionModel-administration.component';
+import { EntitiesAdministrationComponent } from './entity/entities-administration.component';
+import { StatusesAdministrationComponent } from './status/statuses-administration.component';
+import { StatusAdministrationComponent } from './status/status-administration.component';
+import { ActionsAdministrationComponent } from './action/actions-administration.component';
+import { ActionAdministrationComponent } from './action/action-administration.component';
+import { ParameterAdministrationComponent } from './parameter/parameter-administration.component';
+import { ParametersAdministrationComponent } from './parameter/parameters-administration.component';
+import { PrioritiesAdministrationComponent } from './priority/priorities-administration.component';
+import { PriorityAdministrationComponent } from './priority/priority-administration.component';
+import { NotificationsAdministrationComponent } from './notification/notifications-administration.component';
+import { NotificationAdministrationComponent } from './notification/notification-administration.component';
+import { HistoryAdministrationComponent } from './history/history-administration.component';
+import { HistoryBatchAdministrationComponent } from './history/batch/history-batch-administration.component';
+import { UpdateStatusAdministrationComponent } from './updateStatus/update-status-administration.component';
+import { ContactsGroupsAdministrationComponent } from './contact/group/contacts-groups-administration.component';
+import { ContactsGroupAdministrationComponent } from './contact/group/contacts-group-administration.component';
+import { ContactsParametersAdministrationComponent } from './contact/parameter/contacts-parameters-administration.component';
+import { VersionsUpdateAdministrationComponent } from './versionUpdate/versions-update-administration.component';
+import { DocserversAdministrationComponent } from './docserver/docservers-administration.component';
+import { DocserverAdministrationComponent } from './docserver/docserver-administration.component';
+import { TemplatesAdministrationComponent } from './template/templates-administration.component';
+import { TemplateAdministrationComponent } from './template/template-administration.component';
+import { SecuritiesAdministrationComponent } from './security/securities-administration.component';
+import { SendmailAdministrationComponent } from './sendmail/sendmail-administration.component';
+import { ShippingsAdministrationComponent } from './shipping/shippings-administration.component';
+import { ShippingAdministrationComponent } from './shipping/shipping-administration.component';
+import { CustomFieldsAdministrationComponent } from './customField/custom-fields-administration.component';
+import { AppGuard } from '../../service/app.guard';
+import { IndexingModelAdministrationComponent } from './indexingModel/indexing-model-administration.component';
+import { IndexingModelsAdministrationComponent } from './indexingModel/indexing-models-administration.component';
+import { ContactsListAdministrationComponent } from './contact/list/contacts-list-administration.component';
 import { ContactsCustomFieldsAdministrationComponent } from './contact/customField/contacts-custom-fields-administration.component';
 import { ContactsPageAdministrationComponent } from './contact/page/contacts-page-administration.component';
 import { TagsAdministrationComponent } from './tag/tags-administration.component';
 import { TagAdministrationComponent } from './tag/tag-administration.component';
+import { AlfrescoAdministrationComponent } from './alfresco/alfresco-administration.component';
+import { AlfrescoListAdministrationComponent } from './alfresco/alfresco-list-administration.component';
 
 @NgModule({
     imports: [
@@ -71,9 +73,9 @@ import { TagAdministrationComponent } from './tag/tag-administration.component';
             { path: 'administration/parameters', canActivate: [AppGuard], component: ParametersAdministrationComponent },
             { path: 'administration/parameters/new', canActivate: [AppGuard], component: ParameterAdministrationComponent },
             { path: 'administration/parameters/:id', canActivate: [AppGuard], component: ParameterAdministrationComponent },
-            { path: 'administration/priorities', canActivate: [AppGuard], component : PrioritiesAdministrationComponent },
-            { path: 'administration/priorities/new', canActivate: [AppGuard], component : PriorityAdministrationComponent },
-            { path: 'administration/priorities/:id', canActivate: [AppGuard], component : PriorityAdministrationComponent },
+            { path: 'administration/priorities', canActivate: [AppGuard], component: PrioritiesAdministrationComponent },
+            { path: 'administration/priorities/new', canActivate: [AppGuard], component: PriorityAdministrationComponent },
+            { path: 'administration/priorities/:id', canActivate: [AppGuard], component: PriorityAdministrationComponent },
             { path: 'administration/actions', canActivate: [AppGuard], component: ActionsAdministrationComponent },
             { path: 'administration/actions/new', canActivate: [AppGuard], component: ActionAdministrationComponent },
             { path: 'administration/actions/:id', canActivate: [AppGuard], component: ActionAdministrationComponent },
@@ -110,10 +112,13 @@ import { TagAdministrationComponent } from './tag/tag-administration.component';
             { path: 'administration/tags', canActivate: [AppGuard], component: TagsAdministrationComponent },
             { path: 'administration/tags/new', canActivate: [AppGuard], component: TagAdministrationComponent },
             { path: 'administration/tags/:id', canActivate: [AppGuard], component: TagAdministrationComponent },
+            { path: 'administration/alfresco', canActivate: [AppGuard], component: AlfrescoListAdministrationComponent },
+            { path: 'administration/alfresco/new', canActivate: [AppGuard], component: AlfrescoAdministrationComponent },
+            { path: 'administration/alfresco/:id', canActivate: [AppGuard], component: AlfrescoAdministrationComponent },
         ]),
     ],
     exports: [
         RouterModule
     ]
 })
-export class AdministrationRoutingModule {}
+export class AdministrationRoutingModule { }
