@@ -1674,7 +1674,19 @@ export const LANG_EN = {
     "accountLocked": "Too many connections attemps. Retry in",
     "modelUsedByResources": "This model is used by resources, you can't delete it.",
     "mustChangePassword": "Please, you must change your password.",
-    "linkedResources": "Attachments (linked mails)",
+    "linkedResources": "Linked mails",
+    "linkedResourcesAttachments": "Attachments (linked mails)",
     "accessNotFound": "Access not found",
     "moreOneCustom": "This url is not an available instance of this application, please verify your address.",
+    "linkedEntities": "Linked entities",
+    "accounts": "Accounts",
+    "alfrescoCreation": "Alfresco account creation",
+    "alfrescoModification": "Alfresco account modification",
+    "showTreeMsg": "Click here to show the Alfresco tree.",
+    "mustFillAccountMsg": "Please fill the Alfresco account information in order to choose a root folder.",
+    "accountUpdated": "Account updated",
+    "alfrescoAccount": "Alfresco account",
+    "rootFolder": "Root folder",
+    "chooseEntityAssociationAccount": "Choose entities who will be able to use this account",
+    "alfrescoAPI": "API Alfresco address",
 };
