@@ -1673,8 +1673,22 @@ export const LANG_FR = {
     "accountSuspended": "Votre compte utilisateur a été suspendu",
     "accountLocked": "Nombre de tentatives de connexion dépassée. Réessayez dans",
     "modelUsedByResources": "Le modèle est utilisé par des courriers, vous ne pouvez pas le supprimer.",
-    "mustChangePassword": "Vous êtes invité à changer votre mot de passe.",
-    "linkedResources": "Pièces jointes (courriers liés)",
+    "mustChangePassword": "Vous êtes invités à changer votre mot de passe.",
+    "linkedResources": "Courriers liés",
+    "linkedResourcesAttachments": "Pièces jointes (courriers liés)",
     "accessNotFound": "Accès introuvable",
     "moreOneCustom": "Cette url ne correspond à aucune instance configurée, veuillez vérifier l'adresse.",
+    "alfresco": "Alfresco",
+    "adminAlfrescoDesc": "Administrer les comptes alfresco liés aux entités",
+    "linkedEntities": "Entités associées",
+    "accounts": "Comptes",
+    "alfrescoCreation": "Création d'un compte Alfresco",
+    "alfrescoModification": "Modification du compte Alfresco",
+    "showTreeMsg": "Cliquer ici pour afficher l'arborescence du dépôt.",
+    "mustFillAccountMsg": "Veuillez renseigner le compte alfresco pour choisir un dépôt racine.",
+    "accountUpdated": "Compte modifié",
+    "alfrescoAccount": "Compte Alfresco",
+    "rootFolder": "Dépôt racine",
+    "chooseEntityAssociationAccount": "Attribuer le compte au(x) service(s)",
+    "alfrescoAPI": "Adresse de l'API Alfresco",
 };
