@@ -1,6 +1,6 @@
 
 import { Injectable } from '@angular/core';
-import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, CanDeactivate } from '@angular/router';
+import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, CanDeactivate, UrlTree } from '@angular/router';
 import { HttpClient } from '@angular/common/http';
 import { Observable, of } from 'rxjs';
 import { map, tap, catchError, exhaustMap, filter } from 'rxjs/operators';
@@ -9,6 +9,7 @@ import { ProcessComponent } from '../app/process/process.component';
 import { PrivilegeService } from './privileges.service';
 import { AuthService } from './auth.service';
 import { LocalStorageService } from './local-storage.service';
+import { FunctionsService } from './functions.service';
 
 @Injectable({
     providedIn: 'root'
@@ -20,20 +21,20 @@ export class AppGuard implements CanActivate {
         private router: Router,
         private authService: AuthService,
         private localStorage: LocalStorageService,
+        private functionService: FunctionsService,
         public headerService: HeaderService,
         private privilegeService: PrivilegeService
     ) { }
 
     canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<any> {
-        console.log('guard : ' + route.url.join('/'));
-        console.log(state.url.replace(/^\//g, ""));
-
+        console.clear();
+        console.log('== ROUTE GUARD ==');
         this.headerService.resetSideNavSelection();
 
         let tokenInfo = this.authService.getToken();
 
         if (tokenInfo !== null) {
-            console.log('Token trouvé !');
+            this.functionService.debug('Token trouvé !', route.url.join('/'));
 
             this.authService.setUrl(route.url.join('/'));
             if (this.headerService.user.id === undefined) {
@@ -52,12 +53,11 @@ export class AppGuard implements CanActivate {
                                 preferences: data.preferences,
                                 privileges: data.privileges[0] === 'ALL_PRIVILEGES' ? this.privilegeService.getAllPrivileges() : data.privileges
                             };
-
+                            this.functionService.debug('', route.url.join('/'));
                             this.headerService.nbResourcesFollowed = data.nbFollowedResources;
                             this.privilegeService.resfreshUserShortcuts();
 
                             if (this.headerService.user.status === 'ABS') {
-                                console.log('ABS!');
                                 return this.router.navigate(['/activate-user']);
                             } else {
                                 if (route.url.map((url: any) => url.path).filter((url: any) => ['signatureBook', 'content'].indexOf(url) > -1).length > 0) {
@@ -84,10 +84,9 @@ export class AppGuard implements CanActivate {
                         })
                     );
             } else {
-                console.log('Données user trouvé !');
+                console.log('Données user trouvées !');
 
                 if (this.headerService.user.status === 'ABS') {
-                    console.log('ABS!');
                     this.router.navigate(['/activate-user']);
                     return of(false);
                 } else {
@@ -115,19 +114,20 @@ export class AppGuard implements CanActivate {
                         // this.authService.changeKey = data.changeKey;
                         this.localStorage.setAppSession(data.instanceId);
                         tokenInfo = this.authService.getToken();
-                        console.log(tokenInfo);
 
                         if (tokenInfo !== null) {
                             this.authService.setUrl(route.url.join('/'));
+                            this.functionService.debug('', route.url.join('/'));
                             return tokenInfo;
                         } else {
-                            this.authService.setCachedUrl(state.url.replace(/^\//g, ""));
+                            this.authService.setCachedUrl(state.url.replace(/^\//g, ''));
                             console.log('Aucun token trouvé ! Redirection sur login ...');
                             this.authService.logout(false);
                             return false;
                         }
                     }),
                     filter((info: any) => info !== null),
+                    tap(() => console.log('Récupération données user...')),
                     exhaustMap(() => this.http.get('../rest/currentUser/profile')),
                     map((dataUser: any) => {
                         this.headerService.user = {
@@ -141,11 +141,10 @@ export class AppGuard implements CanActivate {
                             preferences: dataUser.preferences,
                             privileges: dataUser.privileges[0] === 'ALL_PRIVILEGES' ? this.privilegeService.getAllPrivileges() : dataUser.privileges
                         };
-
+                        this.functionService.debug('', route.url.join('/'));
                         this.headerService.nbResourcesFollowed = dataUser.nbFollowedResources;
                         this.privilegeService.resfreshUserShortcuts();
                         if (this.headerService.user.status === 'ABS') {
-                            console.log('ABS!');
                             return this.router.navigate(['/activate-user']);
                         } else {
                             if (route.url.map((url: any) => url.path).filter((url: any) => ['signatureBook', 'content'].indexOf(url) > -1).length > 0) {
@@ -181,8 +180,8 @@ export class AppGuard implements CanActivate {
 export class AfterProcessGuard implements CanDeactivate<ProcessComponent> {
     constructor() { }
 
-    async canDeactivate(component: ProcessComponent): Promise<boolean> {
-        if (!component.isActionEnded() && !component.detailMode) {
+    async canDeactivate(component: ProcessComponent, currentRoute: ActivatedRouteSnapshot, currentState: RouterStateSnapshot, nextState: RouterStateSnapshot): Promise<boolean> {
+        if (nextState.url != '/login' && !component.isActionEnded() && !component.detailMode) {
             component.unlockResource();
         }
 
