@@ -1,18 +1,18 @@
 exports.config = {
-    baseUrl: 'http://127.0.0.1/maarch_courrier_develop/cs_recette',
+    baseUrl: 'http://127.0.0.1/MaarchCourrier',
     seleniumAddress: 'http://localhost:4444/wd/hub',
     specs: [
+        'login-spec.js',
         'index-resource-spec.js',
-        //'login-spec.js',
         //'about-us-spec.js'
     ],
     multiCapabilities: [
-        {
-            'browserName': 'chrome',
-            'chromeOptions': {
-                'args': ["--no-sandbox", "--headless", "--disable-gpu",  "--window-size=1920,1080"]
-            },
-        },
+        // {
+        //     'browserName': 'chrome',
+        //     'chromeOptions': {
+        //         'args': ["--no-sandbox", "--headless", "--disable-gpu",  "--window-size=1920,1080"]
+        //     },
+        // },
         {
             'browserName': 'firefox',
             'moz:firefoxOptions': {
